Split the numeric & cast spelling concern out of CSharpPrinter (printer decomposition 2/3)
Slice 2 of the CSharpPrinter decomposition. Moves the cohesive
"numeric reinterpretation & cast spelling" cluster — the last-mile
decisions C# requires but IL leaves implicit — into a new partial-class
file CSharpPrinter.Numerics.cs:

  BinaryText, ShiftCount, ShiftWidthMask, ComparisonText (both overloads),
  IsFloatComparison, UnsignedOperand, CastValue, NumericConstant,
  EffectiveType, BinaryOperator, ComparisonOperator.

Unlike slice 1's name utilities (pure functions of string inputs), these
methods are coupled to the printer's instance state (_fixedLocals,
_function.TypeShapes) and call its recursive Operand/Expression/TypeText,
so the honest tool is a partial-class file split, not a pure static
extraction. The pure type-level decisions are already factored into
TypeFamilies. CSharpPrinter becomes `partial`; the main file is now the
tree-walk visitor.

Verbatim move — zero behavior change. Printer 2150 -> 1967 LOC.
348/348 decompiler tests green.
